Basic persistence of node layout #8
Added simple ontology (layout.owl) to define layout properties. This is used to store the layout for each node in the repository.
Restore the layout when reading in resource map.
Tweaks to ResourceFigure to ensure that plus/minus icon is correctly determined when restoring layout.

Also changed default creator to Anonymous