Merged annotations and compound objects sources into a single tree with hidden root node.
Some minor formatting changes.

Added modified copy of PropertyGrid to ext/ux with enabled header menu - this menu wasn't disabled in the previous version of Ext. Probably better to create our own custom Grid + Column Model eventually, but this is the quickest way to get the header menus for adding/removing metadata back for the demo.