#44 - previous attempt to fix using custom PropertyGrid only worked for active tab - could not track down the reason why - removing PropertyGrid.
Instead, add a toolbar to grids and attach the menus to the buttons.
This is probably more obvious to the user (hidden menu was always an issue).
This approach only works for the non-active tab, so don't activate any tabs by default.

Reorganised init.js a little.

Fixed activation of node properties tab in SelectionProperties