Fix for trac ticket #249, unable to save formatting on annotations.
A contentEditable in firefox defaults to applying formatting as css.

This causes troubles when annotations are saved and reloaded, since the css is stripped out for security reasons.

The Ext HtmlEditor control is supposed to change this 'styleWithCSS' setting, but for some reason it wasn't sticking. This override applies the setting before each command is executed on the contentEditable.