#215 replace compound object tree views with data view that supports paging
The CO manager was changed to use PagingJsonStores instead of CompoundObjectSummary objects so that we can use a DataView directly on the stores. The custom tree views have been updated to work with these stores for testing during development, and I'm committing these changes in case they need to to be used for debugging.

Functionality from tree views (eg context menus, drag and drop etc) has been replicated in the dataview nodes.